handled no authorization exception

+++ WikiConcert/Controllers/BandController.cs
@@ -14,7 +14,15 @@ namespace WikiConcert.Controllers
     {
         private BandService CreateBandService()
         {
-            var userId = Guid.Parse(User.Identity.GetUserId());
+            Guid userId;
+            try
+            {
+                userId = Guid.Parse(User.Identity.GetUserId());
+            }
+            catch (System.ArgumentNullException anex)
+            {
+                return null;
+            }
             var bandService = new BandService(userId);
             return bandService;
         }
@@ -22,6 +30,8 @@ namespace WikiConcert.Controllers
         public IHttpActionResult Get()
         {
             BandService bandService = CreateBandService();
+            if (bandService == null)
+                return Unauthorized();
             var bands = bandService.GetAllBands();
             return Ok(bands);
         }
@@ -29,6 +39,8 @@ namespace WikiConcert.Controllers
         public IHttpActionResult Get(int id)
         {
             BandService bandService = CreateBandService();
+            if (bandService == null)
+                return Unauthorized();
             var bands = bandService.GetBandById(id);
             return Ok(bands);
         }
@@ -37,6 +49,8 @@ namespace WikiConcert.Controllers
         public IHttpActionResult Get(string name)
         {
             BandService bandService = CreateBandService();
+            if (bandService == null)
+                return Unauthorized();
             var bands = bandService.GetBandByName(name);
             return Ok(bands);
         }
@@ -44,6 +58,8 @@ namespace WikiConcert.Controllers
         public IHttpActionResult GetByGenre(string genre)
         {
             BandService bandService = CreateBandService();
+            if (bandService == null)
+                return Unauthorized();
             var bands = bandService.GetBandByGenre(genre);
             return Ok(bands);
         }
@@ -51,6 +67,8 @@ namespace WikiConcert.Controllers
         public IHttpActionResult GetByActive()
         {
             BandService bandService = CreateBandService();
+            if (bandService == null)
+                return Unauthorized();
             var bands = bandService.GetBandByActive(true);
             return Ok(bands);
         }
@@ -58,6 +76,8 @@ namespace WikiConcert.Controllers
         public IHttpActionResult GetByNotActive()
         {
             BandService bandService = CreateBandService();
+            if (bandService == null)
+                return Unauthorized();
             var bands = bandService.GetBandByActive(false);
             return Ok(bands);
         }
@@ -70,8 +90,10 @@ namespace WikiConcert.Controllers
                 return BadRequest(ModelState);
 
             var bandService = CreateBandService();
+            if (bandService == null)
+                return Unauthorized();
 
-            if(bandService.CreateBand(band))
+            if (bandService.CreateBand(band))
                 return Ok($"Successfully added {band.Name}");
 
             return InternalServerError();
@@ -85,6 +107,8 @@ namespace WikiConcert.Controllers
                 return BadRequest(ModelState);
 
             var bandService = CreateBandService();
+            if (bandService == null)
+                return Unauthorized();
 
             if (bandService.UpdateBand(band))
                 return Ok($"Successfully updated {band.Name}");
@@ -95,6 +119,8 @@ namespace WikiConcert.Controllers
         public IHttpActionResult BandDelete(int id)
         {
             var bandService = CreateBandService();
+            if (bandService == null)
+                return Unauthorized();
 
             if (bandService.DeleteBand(id))
                 return Ok("Successfully deleted band.");
